requirements.txt
alive-progress==2.4.1
dnspython==2.2.1
ImageHash==4.2.1
numpy==1.22.3
Pillow==9.1.0
pymongo==4.1.1
python-telegram-bot==20.0a0
requests==2.27.1
selenium==4.1.5
vk==2.0.2
vk-api==11.9.8
python-dotenv==0.20.0